Canonicalize.
